+++ apps/pages/forms.py
@@ -8,6 +8,7 @@ class Page_translationInlineForm(ModelForm):
         model = Page_translation
         exclude = ['updated_at', 'created_at']
 
+
 class PageForm(ModelForm):
 
     def __init__(self, *args, **kwargs):
@@ -19,21 +20,7 @@ class PageForm(ModelForm):
         exclude = ['updated_at', 'created_at']
 
     class Media:
-        js = ('jscolor/jscolor.js',)
+        js = ('vendor/jscolor/jscolor.js',)
         #css = {
             #'all': ('zzzZZZzzz.css',)
         #}
-
-    # def clean_photo(self):
-    #     new_photo = self.cleaned_data['photo']
-    #     self.
-    #     if new_photo != self.old_photo:
-    #         from PIL import Image
-    #         path = new_photo
-    #         image = Image.open(path)
-    #         image.save(path,quality=20,optimize=True)
-    #         print('PHOTO UPDATE')
-    #     else:
-    #         print('NOT_NOT-NOT')
-    #
-    #     return new_photo