+++ config/local.py
@@ -6,7 +6,7 @@ DEBUG = True
 
 COMPRESS_ENABLED = False
 
-DEBUG_TOOLBAR_PATCH_SETTINGS = False
+DEBUG_TOOLBAR_PATCH_SETTINGS = True
 
 SECRET_KEY = '=)1+x0_f%6t&$4gftv-^^w20kr)+1ut8js0@0(gt6cx8z6kxlc'
 
@@ -32,7 +32,8 @@ ALLOWED_HOSTS = ['127.0.0.1', 'localhost']
 MEDIA_ROOT = os.path.join(BASE_DIR, 'media/')
 
 # Enter path to static folder
-STATIC_ROOT = ''
+STATIC_ROOT = os.path.join(BASE_DIR, 'static/')
+
 
 COMPRESS_ROOT = os.path.join(BASE_DIR, 'static/')