+++ config/mysmile.py
@@ -1,7 +1,7 @@
 # MySmile settings
 from django.conf import settings
 
-MYSMILE_VERSION='0.6.1'
+MYSMILE_VERSION = '0.6.2'
 
 # for MYSMILE_VERSION availability in templates
 TEMPLATE_CONTEXT_PROCESSORS = settings.TEMPLATE_CONTEXT_PROCESSORS + (